# Organization/Organization.Client/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Organization.Server",
    "version": "1.0"
  },
  "paths": {
    "/api/Department": {
      "get": {
        "tags": [
          "Department"
        ],
        "summary": "The method returns all the departments in the organization",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetDepartmentDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetDepartmentDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetDepartmentDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Department"
        ],
        "summary": "The method adds a new department into organization",
        "requestBody": {
          "description": "A new department that need to be added",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostDepartmentDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PostDepartmentDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PostDepartmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetDepartmentDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetDepartmentDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetDepartmentDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Department/{id}": {
      "get": {
        "tags": [
          "Department"
        ],
        "summary": "The method returns a department by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Department ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetDepartmentDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetDepartmentDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetDepartmentDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Department"
        ],
        "summary": "The method updates a department information by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "An ID of the department",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "New information of the department",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostDepartmentDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PostDepartmentDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PostDepartmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetDepartmentDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetDepartmentDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetDepartmentDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Department"
        ],
        "summary": "The method deletes a department by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "An ID of the department",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetDepartmentDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetDepartmentDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetDepartmentDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GetDepartmentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Id - an id of the department",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "description": "Name - a name of the department",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "class GetDepartmentDto - represents a department in company"
      },
...
      "PostDepartmentDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name - a name of the department",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "class PostDepartmentDto - represents a department in company"
      },
...
    }
  }
}